MA-14166 Coverity 3623702: NEGATIVE_RETURNS in Component of opencl-2d.cpp
Mandatory type conversion of variable that may return negative values.

Convert the return variable of ftell to size_t, as the same as
program_length.

Reviewed-on: http://androidsource.ap.freescale.net/project/8161

+++ opencl-2d/opencl-2d.cpp
@@ -348,7 +348,7 @@ static cl_program CreateProgram(cl_context context, cl_device_id device,
 
     // get the length of the source code
     fseek(pFileStream, 0, SEEK_END);
-    program_length = ftell(pFileStream);
+    program_length = (size_t)ftell(pFileStream);
     fseek(pFileStream, 0, SEEK_SET);
 
     // allocate a buffer for the source code string and read it in
